Make Like non-subclassable concrete (#649)
